New test voor BENSON-17
